+++ R/R6Class-partlyObservedNetwork.R
@@ -79,11 +79,7 @@ partlyObservedNetwork <-
       private$phi <- covar$Array
 
       ## sets of observed / unobserved dyads
-      if (private$directed) {
-        dyads <- .row(dim(adjacencyMatrix)) != .col(dim(adjacencyMatrix))
-      } else {
-        dyads <- .row(dim(adjacencyMatrix)) < .col(dim(adjacencyMatrix))
-      }
+      dyads <- valid_dyads(dim(adjacencyMatrix), private$directed)
       ## where are my observations?
       NAs   <- is.na(adjacencyMatrix)
       obs   <- which(!NAs & dyads, arr.ind = TRUE )
@@ -118,7 +114,7 @@ partlyObservedNetwork <-
       d <- 1/sqrt(rowSums(abs(A)))
       D <- Diagonal(x = d)
       U <- eigs_sym(D %*% A %*% D, max(vBlocks))$vectors
-      res <- future_lapply(vBlocks, function(k) {
+      res <- future_lapply_shuffled(vBlocks, function(k) {
         cl <- rep(1L, n)
         if (k != 1) {
           Un <- U[, 1:k, drop = FALSE]
@@ -127,12 +123,14 @@ partlyObservedNetwork <-
           cl_ <- as.integer(
             kmeans_missSBM(Un, k)
           )
-         ## handing lonely souls
+         ## handling lonely souls: nothing connects them to the rest of the graph, so there is
+         ## no basis to pick a class for them individually -- put them all in whichever class
+         ## carries the most spectral weight, i.e. the most "typical" one
          cl[connected] <- cl_
          cl[unconnected] <- which.max(rowsum(d, cl_))
         }
         cl
-      }, future.seed = TRUE, future.scheduling = structure(TRUE, ordering = "random"))
+      })
       res
     },
     #' @description basic imputation from existing clustering